server/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.35,<1
python-dotenv>=1.1,<2
python-multipart>=0.0.20,<1
faiss-cpu>=1.9,<2
sentence-transformers>=3.2,<6
numpy>=1.26,<3
requests>=2.32,<3
pypdf>=5.4,<6
psycopg[binary]>=3.2,<4
pgvector>=0.3,<1
joblib>=1.5,<2
scikit-learn>=1.6,<2
nltk>=3.9,<4
